pyproject.toml
[tool.poetry]
name = "sd2-depth-api"
version = "0.1.0"
description = ""
authors = ["Jiri Koutny"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "1.1.4"
colorlog = "^6.7.0"
diffusers = "^0.11.1"
transformers = { git = "https://github.com/huggingface/transformers", rev = "092d4d49dd556d18598cc541ce2fc3477822398f" }
accelerate = "^0.15.0"
scipy = "^1.9.3"
safetensors = "^0.2.7"
Pillow = "^9.3.0"
logzio-python-handler = "3.1.1"
opencv-python = "^4.7.0"
matplotlib = "^3.6.3"
boto3 = "^1.26.57"
torchvision = "0.14.0"
numpy = "1.23.4"
imutils = "0.5.4"
timm = "0.6.12"
einops = "0.6.0"

[tool.poetry.dev-dependencies]
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
